refactor(Step): move pacakge
