+++ src/main/java/nextstep/ladder/domain/step/strategy/RandomStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain.step;
+package nextstep.ladder.domain.step.strategy;
 
 import java.util.Random;
